[RF] Move RooDataWeightedAverage to RooFitLegacy
