+++ roofit/roofitcore/inc/RooAbsReal.h
@@ -136,7 +136,7 @@ public:
   virtual double getValV(const RooArgSet* normalisationSet = nullptr) const ;
 
   virtual RooSpan<const double> getValues(RooBatchCompute::RunContext& evalData, const RooArgSet* normSet = nullptr) const;
-  std::vector<double> getValues(RooAbsData const& data, const RooArgSet* normSet = nullptr) const;
+  std::vector<double> getValues(RooAbsData const& data) const;
 
   double getPropagatedError(const RooFitResult &fr, const RooArgSet &nset = RooArgSet()) const;
