+++ roofit/roofitcore/inc/RooFitLegacy/RooDataWeightedAverage.h
